Remove outdated tads2/glk implementation Embed osfrobtads.h into osglk.h as it was confusing to have essential code in a file for another port
