--- tads2/glk/glkver.txt
@@ -1,41 +1,0 @@
-Recent changes to GlkTADS
---------------------------
-
-This file contains a list of recent changes to TADS that are specific
-to the Glk libraries.
-
-Note that a separate file, called TADSVER.HTM, contains the release
-notes apply to all types of computers, not only Glk.  Please refer to
-TADSVER.HTM for additional information about this release.
-
-
-------------------------------------------------------------------------------
-2.4.0/Glk version 1
-
-  - A CheapGlk port has been added, and the Linux port settings have
-    been changed to account for different distributions.
-
-  - The Makefile now takes the OPTIMIZE flags into account.
-
-  - The paths hack in the Unix port has been extensively reworked. When
-    GlkTADS tries to open a file, it first looks to see if there is a
-    path associated with the filename. If so, it changes to that
-    directory. The file is opened; if the directory was changed, GlkTADS
-    changes the directory back to the original directory. All of this
-    is done in the oss_check_path and oss_revert_path functions in
-    oss_glk.c, if you wish to change this behavior.
-
-  - GlkTADS used to add an extension to saved games whether or not one
-    already existed, a boon to people who like their saved games to be
-    "s20.sav.sav". For those of us who don't, however, this has been
-    fixed.
-
-
-------------------------------------------------------------------------------
-2.3.0/Glk version 1
-
-  - Initial version using Glk library 0.5. Versions exist for GlkTerm
-    and XGlk.
-
-  - Note that, under GlkTerm, you can access the scrollback using
-    Ctrl-V and Ctrl-Y. This is part of GlkTerm, not part of GlkTADS.
